backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
tavily-python
sentence-transformers
langsmith
python-dotenv
arxiv
wikipedia
google-api-python-client
pydantic
